feat(remote): making remote2 primary remote
- delete the original `remote` cmd
- promote the new `remote2` cmd to be the main `remote` cmd
- remove references to `remote2`
